+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 ### fixed
 * evaluate 类的可视化函数补充 rendered 选项
+* 修正了 tradevolume 周数和 pandas 给出的 iso 周数不一定对应的问题
 
 ## v0.9.3 - 2020.06.20
 ### added